+++ asir2000/include/ca.h
@@ -45,7 +45,7 @@
  * DEVELOPER SHALL HAVE NO LIABILITY IN CONNECTION WITH THE USE,
  * PERFORMANCE OR NON-PERFORMANCE OF THE SOFTWARE.
  *
- * $OpenXM: OpenXM_contrib2/asir2000/include/ca.h,v 1.25 2001/09/13 03:04:28 noro Exp $ 
+ * $OpenXM: OpenXM_contrib2/asir2000/include/ca.h,v 1.26 2001/09/17 07:16:59 noro Exp $ 
 */
 #include <stdio.h>
 
@@ -374,7 +374,7 @@ typedef struct oDL {
 typedef struct oCDP {
 	int len;
 	int psindex;
-	unsigned short *body;
+	unsigned int *body;
 } *CDP;
 
 typedef struct oCM {
